// src/components/Organisms/AssetTile/AssetTile.interface.ts
export enum AssetTileLayout {
  auto = 'auto',
  listItem = 'listItem',
  gridItem = 'gridItem',
  compact = 'compact',
}

/** AssetTileOrientation `landscape | portrait` use portrait on mobile and landscape on desktop */
export type AssetTileOrientation = 'landscape' | 'portrait' | undefined

/** AssetTileSize  use small on mobile and large on desktop */
export type AssetTileSize = 'auto' | 'large' | 'small'

export type AssetView = 'list' | 'collection' | 'compact' | 'sponsored' | 'disabled'

export type playlistView = 'list' | 'collection' | 'disabled'

export enum AssetStrategy {
  structured = '1',
  timebox = '2',
  none = 'none',
}

export interface SponsoringDetails {
  isSponsored: boolean
  showUppercase: boolean
  label: string
  infoText: string
  title: string
}

export interface CollectionData extends Partial<Asset> {
  totalAssetDuration: number | undefined
  completionPercentage: number | undefined
  numberOfAssets: number | undefined
  title: string
  pid?: string | undefined
}

export interface AssetTileProps {
  /**  image padding for asset images */
  badgeActionType?: BadgeActionType
  badges?: (BadgeType | null)[]
  borderColor?: ThemeColors
  fullHeight?: boolean
  isCompact?: boolean
  expanded?: boolean
  isOpenAsset?: boolean
  eventDate?: string
  isCollection?: boolean
  placeCard?: boolean
  assetIndex?: number
  collectionData?: CollectionData
  assetSponsoring?: SponsoringDetails | undefined
  assetView?: AssetView
  headerFontSize?: string
  strategyType?: AssetStrategy | undefined
  color?: ThemeColors
  elevation?: Elevation
  iconsRight?: boolean
  inView?: boolean
  elevationHover?: Elevation
  noBorder?: CardNoBorder
  headlineLimitLines?: number
  limitContentLines?: number
  roundImages?: boolean
  icon?: React.ReactElement
  lazyloadLowQuality?: boolean
  orientation?: AssetTileOrientation
  competencyLabel?: string
  size?: AssetTileSize
  fontSize?: ThemeFontSizes
  mainLink?: any
  isDisabled?: boolean
  isMobileView?: boolean
  layout?: AssetTileLayout
  lazyLoad?: boolean
  loading?: boolean
  loadingOperation?: boolean
  locked?: boolean
  showBadges?: boolean
  showLock?: boolean
  showBookmark?: boolean
  showProvider?: boolean
  showFeatured?: boolean
  showLikes?: boolean
  showInteractions?: boolean
  showOptions?: boolean
  showSkills?: boolean
  showImage?: boolean
  showComplete?: boolean
  showAssetMeta?: boolean
  showMoreText?: string
  showLessText?: string
  showMoreNumOfLines?: number
  showMoreAlignment?: FlexAlignItemsOptions
  fadeOutColor?: string
  videoFallbackImage?: string
  renderAddToBookmarkButton?: (AssetForBookmarking: {
    id?: number | undefined
    bookmarked?: boolean
    title?: string
    provider?: Provider | undefined
    duration?: string | undefined
    type?: AssetType
    completed?: any
    assetIndex?: number
    playlistId?: string | undefined
  }) => JSX.Element | React.ReactNode
  onHeadingClick?: (e: React.MouseEvent<HTMLDivElement>) => void
  onOptionsClick?: (e: React.MouseEvent<HTMLDivElement>) => void
  onClick?: (event: React.MouseEvent<HTMLElement, MouseEvent>) => void
  onPictureClick?: (event: React.MouseEvent<HTMLElement, MouseEvent>) => void
  onLockClick?: (event: React.MouseEvent<HTMLElement, MouseEvent>) => void | undefined
  onCompleteClick?: ((ev: React.MouseEvent<HTMLElement, MouseEvent>) => void) | undefined
  asset: Asset | Partial<Asset>
  sponsoringDetails?: SponsoringDetails
  withEllipsis?: boolean
  withLQIP?: boolean
  withNativeLoading?: boolean
}

export interface StyledResponsiveContainerProps {
  orientation: AssetTileOrientation
  loading?: boolean
  isSmall?: boolean
}
export interface StyledDivider {
  height?: string
}

export interface StyledTileProps {
  orientation: AssetTileOrientation
}

export interface ContentTextProps {
  contentText?: string
  loading: boolean
  isDisabled?: boolean
  limitContentLines?: number
}

export interface TileSettings {
  showAssetMeta?: boolean
  sponsoringDetails?: undefined
  showLock?: boolean
  showSkills?: boolean
  withEllipsis: boolean
  headlineLimitLines: number
  fontSize?: string
  loading: boolean | undefined
  showBookmark?: boolean | undefined
  showFeatured?: boolean | undefined
}
// ...
export type Provider = {
  __typename: string
  id: number
  name: string
  logoUrl?: string
}